More work on moving all to the new MJ api

+++ src/api/java/buildcraft/api/fuels/IFuelManager.java
@@ -11,7 +11,7 @@ import net.minecraftforge.fluids.Fluid;
 public interface IFuelManager {
     IFuel addFuel(IFuel fuel);
 
-    IFuel addFuel(Fluid fluid, int powerPerCycle, int totalBurningTime);
+    IFuel addFuel(Fluid fluid, double powerPerCycle, int totalBurningTime);
 
     Collection<IFuel> getFuels();
