+++ src/energy/java/buildcraft/energy/fuels/FuelManager.java
@@ -27,7 +27,7 @@ public final class FuelManager implements IFuelManager {
     }
 
     @Override
-    public IFuel addFuel(Fluid fluid, int powerPerCycle, int totalBurningTime) {
+    public IFuel addFuel(Fluid fluid, double powerPerCycle, int totalBurningTime) {
         return addFuel(new BCFuel(fluid, powerPerCycle, totalBurningTime));
     }
 
@@ -48,10 +48,10 @@ public final class FuelManager implements IFuelManager {
 
     private static final class BCFuel implements IFuel {
         private final Fluid fluid;
-        private final int powerPerCycle;
+        private final double powerPerCycle;
         private final int totalBurningTime;
 
-        public BCFuel(Fluid fluid, int powerPerCycle, int totalBurningTime) {
+        public BCFuel(Fluid fluid, double powerPerCycle, int totalBurningTime) {
             this.fluid = fluid;
             this.powerPerCycle = powerPerCycle;
             this.totalBurningTime = totalBurningTime;
@@ -68,7 +68,7 @@ public final class FuelManager implements IFuelManager {
         }
 
         @Override
-        public int getPowerPerCycle() {
+        public double getPowerPerCycle() {
             return powerPerCycle;
         }
     }
